(feat): updated FirstRunStorage
