+++ lib/persistence/storage/first_run/i_first_run_storage.dart
@@ -1,0 +1,8 @@
+/// Repository interface for working with first app status
+abstract class IFirstRunStorage {
+  /// Returns is it first app run
+  bool getIsFirstRun();
+
+  /// Save first run app
+  Future<void> setIsFirstRun({required bool value});
+}
